Fix build: bindgen sysroot, clippy warnings, no_std correctness
build.rs:
- Fix bindgen use_core() call (v0.70 takes no argument)
- Add gcc_builtin_include() helper to locate stddef.h/stdint.h for
  libclang when it cannot find system headers automatically

src/packet.rs: std::mem/ffi -> core::mem/ffi for no_std compatibility
src/route.rs: import alloc::vec macro explicitly for no_std; Vec import
src/connection.rs: remove redundant csp_read cast (clippy)
src/socket.rs: remove redundant csp_recvfrom cast (clippy)
src/init.rs: remove redundant u32 cast on ping; suppress too_many_args
  on transaction() which intentionally mirrors the C API signature

Verified: cargo build, cargo clippy (clean), cargo check --no-default-features

+++ src/init.rs
@@ -254,6 +254,7 @@ impl CspNode {
     // ── One-shot transactions ─────────────────────────────────────────────
 
     /// Perform a full request/reply transaction (new connection each call).
+    #[allow(clippy::too_many_arguments)]
     ///
     /// Creates a connection, sends `out_buf`, waits for a reply into `in_buf`,
     /// then closes the connection.
@@ -322,7 +323,7 @@ impl CspNode {
 
     /// Send a ping to `node` and return the echo time in ms, or -1 on error.
     pub fn ping(&self, node: u8, timeout: u32, payload_size: u32, opts: u8) -> i32 {
-        unsafe { sys::csp_ping(node, timeout, payload_size as u32, opts) }
+        unsafe { sys::csp_ping(node, timeout, payload_size, opts) }
     }
 
     /// Request and return free memory on `node`.
